requirements.txt
APScheduler==3.8.1
boto3==1.19.10
Django==3.1.3
django-cacheops==6.1
django-cors-headers==3.10.0
django-storages==1.12.3
dj_static==0.0.6
djangorestframework==3.12.4
drf-yasg==1.20.0
elasticsearch==7.17.4
firebase-admin==5.2.0
google-api-python-client==2.58.0
gunicorn==20.1.0
mysql-client==0.0.1
mysqlclient==2.0.3
Pillow==8.4.0
redis==4.3.3
selenium==4.2.0
